[#1851] bootstrap to use the same code and save survey groups

+++ utils/src/main/java/org/akvo/flow/utils/entity/Form.kt
@@ -22,7 +22,7 @@ package org.akvo.flow.utils.entity
 data class Form(
     val id: Int,
     val formId: String,
-    val surveyId: Int,
+    val surveyId: Long,
     val name: String,
     val version: Double,
     val type: String = "survey",
